+++ skrub/_reporting/_sample_table.py
@@ -63,6 +63,10 @@ Table Cell
   rowspan, colspan, scope: the values for the html attributes with the same names
 """
 import pandas as pd
+try:
+    from polars import selectors as cs
+except ImportError:
+    pass
 
 from .. import _dataframe as sbd
 from .._dispatch import dispatch
@@ -93,13 +97,25 @@ def make_table(df, max_top_slice_size=5, max_bottom_slice_size=5):
 
 
 @make_table.specialize("pandas", argument_type="DataFrame")
-def _make_table_pandas(df, max_top_slice_size=5, max_bottom_slice_size=5):
-    return _PandasTable(df, max_top_slice_size, max_bottom_slice_size).table_data
+def _make_table_pandas(
+    df, max_top_slice_size=5, max_bottom_slice_size=5, float_precision=None
+):
+    return _PandasTable(
+        df, max_top_slice_size, max_bottom_slice_size, float_precision
+    ).table_data
 
 
 @make_table.specialize("polars", argument_type="DataFrame")
-def _make_table_polars(df, max_top_slice_size=5, max_bottom_slice_size=5):
-    return _PolarsTable(df, max_top_slice_size, max_bottom_slice_size).table_data
+def _make_table_polars(
+    df, max_top_slice_size=5, max_bottom_slice_size=5, float_precision=None
+):
+    if float_precision is not None and float_precision < 0:
+        raise ValueError(
+            "Negative rounding precision is not supported for Polars."
+        )
+    return _PolarsTable(
+        df, max_top_slice_size, max_bottom_slice_size, float_precision
+    ).table_data
 
 
 def _pick_slice_sizes(df, max_top_size, max_bottom_size):
@@ -120,10 +136,11 @@ def _pick_slice_sizes(df, max_top_size, max_bottom_size):
 class _PolarsTable:
     """Helper for _make_table_polars."""
 
-    def __init__(self, df, max_top_slice_size, max_bottom_slice_size):
+    def __init__(self, df, max_top_slice_size, max_bottom_slice_size, float_precision):
         self.df = df
         self.max_top_slice_size = max_top_slice_size
         self.max_bottom_slice_size = max_bottom_slice_size
+        self.float_precision = float_precision
 
         self.start_i = -1
         self.start_j = 0
@@ -170,6 +187,8 @@ class _PolarsTable:
             )
 
     def add_table_body(self, sub_df, part_name, start_i):
+        if self.float_precision is not None:
+            sub_df = sub_df.with_columns(cs.float().round(self.float_precision))
         tbody = {"name": part_name, "elem": "tbody", "rows": []}
         self.parts.append(tbody)
         for tr_count, df_row in enumerate(sub_df.iter_rows()):
@@ -334,10 +353,11 @@ class _PandasTable:
     stop_i, stop_j are one past the last i, j coords (df.shape[0], df.shape[1])
     """
 
-    def __init__(self, df, max_top_slice_size, max_bottom_slice_size):
+    def __init__(self, df, max_top_slice_size, max_bottom_slice_size, float_precision):
         self.df = df
         self.max_top_slice_size = max_top_slice_size
         self.max_bottom_slice_size = max_bottom_slice_size
+        self.float_precision = float_precision
 
         self.n_col_levels = _n_levels(self.df.columns)
         self.n_idx_levels = _n_levels(self.df.index)
@@ -489,6 +509,8 @@ class _PandasTable:
             )
 
     def add_table_body(self, sub_df, part_name, start_i):
+        if self.float_precision is not None:
+            sub_df = sub_df.round(self.float_precision)
         tbody = {"name": part_name, "elem": "tbody", "rows": []}
         self.parts.append(tbody)